Code and repository refractor
Code and repository changes that complete version 2.0 and remove all the version 1.0 files.
