+++ src/App.css
@@ -1,5 +1,109 @@
-.App{
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
+html, body, section {
+    block-size: 100%;
+  }
+  
+html {
+  scroll-snap-type: y mandatory;
 }
+
+section {
+  scroll-snap-align: start;
+  scroll-snap-stop: always;
+  
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: 28px;
+}
+
+section:nth-of-type(even) {
+  color: white;
+  background: black;
+}
+
+section:nth-of-type(odd) {
+  color: black;
+  background: white;
+}
+
+a{
+  text-decoration: none;
+  color: black;
+}
+
+.link:hover {
+  text-shadow: 2px 2px 5px black;
+}
+
+body {
+  margin: auto;
+  font-family: system-ui, sans-serif;
+  font-size: 2.5rem;
+}
+
+button{
+  text-decoration: none;
+  color: black;
+}
+
+.container-row{
+  
+  display: flex;
+  
+}
+
+.container-col{
+  display: flex;
+  flex-direction: column;
+}
+
+img{
+  max-width: 90%;
+  height: auto;
+}
+
+.item{
+  display: block;
+  text-align: center;
+  width: 25%;
+}
+
+.button1 {
+  background-color: #7327ca; 
+  border: none;
+  color: white;
+  padding: 15px 32px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 16px;
+  
+}
+
+.button2 {
+  background-color:#0072b1; 
+  border: none;
+  color: white;
+  padding: 15px 32px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 16px;
+}
+
+button:hover {
+  box-shadow: 0 12px 16px 0 rgba(0,0,0,0.24), 0 17px 50px 0 rgba(0,0,0,0.19);
+}
+
+.button3 {
+  background-color:#f44336; 
+  border: none;
+  color: white;
+  padding: 15px 32px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 16px;
+}
+
