+++ src/App.js
@@ -1,25 +1,26 @@
 import React from 'react';
-import {BrowserRouter, Switch, Route, Redirect } from 'react-router-dom';
-import './App.css'
-import NavBar from './components/NavBar/NavBar';
-import Home from './components/Home/Home';
+import './App.css';
+import Landing from './components/Landing/Landing';
+import About from './components/About/About';
+import Projects from './components/Projects/Projects';
 import Resume from './components/Resume/Resume';
+import Contact from './components/Contact/Contact';
 
-export default function App() {
-  return (
-    <BrowserRouter> 
-        <Redirect from='/' to='/Home' />
-        <div class='App'>
-          <NavBar />
-          <Switch>
-            <Route path ="/Home">
-              <Home />
-            </Route>
-            <Route path ="/Resume">
-              <Resume />
-            </Route>
-          </Switch>
-        </div>
-    </BrowserRouter> 
-    );
-  }
+export default function Npp(){
+
+    return(
+        <>
+            <Landing />
+            
+            <About />
+            
+            <Projects />
+
+            <Resume />
+
+            <Contact />
+            
+            
+        </>
+    )
+}
