+++ src/components/Contact/Contact.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export default function Contact(){
+    return(
+        <section>
+            <h1>Contact Me</h1>
+            <ul>
+                <li>Email (preferred): parkerhansen7777@gmail.com</li>
+                <li>Phone: 817-404-9343</li>
+            </ul>
+        </section> 
+    )
+}
