--- src/components/Home/Home.css
@@ -1,76 +1,0 @@
-.p-body{
-    flex-grow: 10;
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-    background-color: rgb(61, 79, 184);
-    color: white;
-}
-
-img{
-    width: 15%;
-    height: auto;
-}
-
-.button1 {
-    background-color: #7327ca; 
-    border: none;
-    color: white;
-    padding: 15px 32px;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    font-size: 16px;
-    
-  }
-
-  .button2 {
-    background-color:#f44336; 
-    border: none;
-    color: white;
-    padding: 15px 32px;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    font-size: 16px;
-  }
-
-  .button3 {
-    background-color:#0072b1; 
-    border: none;
-    color: white;
-    padding: 15px 32px;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    font-size: 16px;
-  }
-  a{
-    text-decoration: none;
-    color: white;
-    display: block;
-    text-decoration: underline;
-  }
-
-
-.links{
-    padding-top: 20px;
-}
-
-.Page-footer {
-	margin-top: auto;
-  background: #838383;
-  padding: 10px 10px;
-	text-align: center;
-}
-
-.flex-item{
-  min-width: 25%;
-  margin: auto;
-  gap: 10px;
-}
-
-p{
-  width: 25%;
-  margin: auto;
-}
