+++ src/components/Landing/Landing.js
@@ -1,25 +1,12 @@
 import React from 'react';
-import './Landing.css';
 
-class Landing extends React.Component{
-    render(){
-        return(
-            <div class='p-body'>
-                <section>
-                    <h1>Hello, my name is Parker Hansen...</h1>
-                </section>
-                <section>
-                    <h1>Hello, my name is Parker Hansen...</h1>
-                </section> 
-                <section>
-                    <h1>Hello, my name is Parker Hansen...</h1>
-                </section> 
-                <section>
-                    <h1>Hello, my name is Parker Hansen...</h1>
-                </section>  
-            </div>
-        );
-     }   
-    }
-    
-    export default Landing
+
+export default function Landing(){
+    return(
+        <section>
+            <h1>Hello, my name is Parker Hansen....</h1>
+            <span>Scroll down</span>
+            <span>\/</span>
+        </section>
+    )
+}
