--- src/components/NavBar/NavBar.css
@@ -1,33 +1,0 @@
-
-#navbar {
-	margin: 0;
-	text-align: center;
-	padding: 0;
-	list-style: none;
-	background-color: rgb(55, 85, 253);
-	border-bottom: 3px solid rgb(32, 48, 139); 
-	border-top: 3px solid rgb(32, 48, 139); }
-	
-
-.link {
-	
-	display: inline-block;
-	text-align: center;
-	padding: 4px 50px;
-	text-decoration: none;
-	font-weight: bold;
-	color: rgb(255, 255, 255);
-	
-}
-.link:hover{
-
-	color: rgb(40, 62, 188);
-	background-color: rgb(255, 255, 255);
-
-}
-
-.selected{
-
-	color : goldenrod;
-
-}
