+++ src/components/Resume/Resume.js
@@ -9,13 +9,10 @@ export default function Resume(){
     
     return(
     
-        <div class="p-body">
-            
-            <h1>Welcome to my Resume</h1>
-            <SinglePage pdf={resumePDF} />
-            
-            <footer className="Page-footer"><span>Created by Parker (© 2021)</span></footer>
-        </div>
+        <section>
+                <h1>My Resume</h1>
+                <SinglePage pdf={resumePDF}/>
+            </section>
      );
