+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {createRoot} from 'react-dom/client';
 import './index.css';
 //import App from './App';
-import Npp from './Npp';
+import Npp from './App';
 import * as serviceWorker from './serviceWorker';
 
 const rootElement = document.getElementById('root')
@@ -10,7 +10,6 @@ const root = createRoot(rootElement);
 
 root.render(
   <Npp />
-  //<App />
 );
 
 // If you want your app to work offline and load faster, you can change
